+++ tests/testthat/test-lrren.R
@@ -296,23 +296,23 @@ test_that("lrren works", {
           verbose = FALSE)
   )
 
-  # # Estimate and cross-validation
-  # ## Unbalanced sampling
-  # ## Parallel (n = 2 cores)
-  # expect_named(
-  #   lrren(obs_locs = obs_locs,
-  #         predict = FALSE,
-  #         predict_locs = NULL,
-  #         conserve = TRUE,
-  #         cv = TRUE,
-  #         kfold = 10,
-  #         balance = TRUE,
-  #         parallel = TRUE,
-  #         n_core = 2,
-  #         poly_buffer = NULL,
-  #         obs_window = NULL,
-  #         verbose = FALSE)
-  # )
+  # Estimate and cross-validation
+  ## Unbalanced sampling
+  ## Parallel (n = 2 cores)
+  expect_named(
+    lrren(obs_locs = obs_locs,
+          predict = FALSE,
+          predict_locs = NULL,
+          conserve = TRUE,
+          cv = TRUE,
+          kfold = 10,
+          balance = TRUE,
+          parallel = TRUE,
+          n_core = 2,
+          poly_buffer = NULL,
+          obs_window = NULL,
+          verbose = FALSE)
+  )
 
   # Estimate and predict
   expect_named(